# Add JSON benchmarks for missing-property and case-insensitive