Adds new benchmarks to track changes made in https://github.com/dotnet/runtime/pull/35848

Verifies ~2x perf improvement on the missing-property and case-insensitive scenarios.

Before changes from the runtime PR linked above:
```
|                     Method |     Mean |     Error |    StdDev |   Median |      Min |      Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |---------:|----------:|----------:|---------:|---------:|---------:|-------:|------:|------:|----------:|
|        DeserializeBaseline | 1.209 us | 0.0139 us | 0.0130 us | 1.207 us | 1.186 us | 1.231 us | 0.0676 |     - |     - |     448 B |
|         DeserializeMissing | 1.593 us | 0.0161 us | 0.0150 us | 1.595 us | 1.557 us | 1.622 us | 0.0386 |     - |     - |     264 B |
|    DeserializeCaseMatching | 1.304 us | 0.0120 us | 0.0112 us | 1.301 us | 1.289 us | 1.329 us | 0.1169 |     - |     - |     752 B |
| DeserializeCaseNotMatching | 2.369 us | 0.0240 us | 0.0224 us | 2.371 us | 2.333 us | 2.411 us | 0.1426 |     - |     - |     928 B |
```

After:
```
|                     Method |       Mean |    Error |   StdDev |     Median |        Min |        Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------------------- |-----------:|---------:|---------:|-----------:|-----------:|-----------:|-------:|------:|------:|----------:|
|        DeserializeBaseline | 1,228.1 ns | 12.32 ns | 11.53 ns | 1,226.9 ns | 1,210.2 ns | 1,246.1 ns | 0.0685 |     - |     - |     448 B |
|         DeserializeMissing |   841.5 ns |  6.32 ns |  5.60 ns |   840.0 ns |   831.2 ns |   851.4 ns | 0.0134 |     - |     - |      88 B |
|    DeserializeCaseMatching | 1,224.4 ns | 12.83 ns | 12.00 ns | 1,224.7 ns | 1,209.8 ns | 1,246.2 ns | 0.0676 |     - |     - |     448 B |
| DeserializeCaseNotMatching | 1,234.0 ns |  9.58 ns |  8.49 ns | 1,233.8 ns | 1,223.1 ns | 1,249.1 ns | 0.0688 |     - |     - |     448 B |
```