@billwert 2020-05-12
Thanks @steveharter! LGTM. We'll sort out the comparison tool separately.